@@ -186,25 +197,27 @@
     "package-change-checker": "^1.2.4",
     "prettier": "2.2.1",
     "pretty-proptypes": "^0.6.0",
+    "process": "^0.11.10",
     "raw-loader": "^4.0.2",
     "react": "18",
     "react-codesandboxer": "^3.1.1",
     "react-dom": "18",
     "react-helmet": "^5.2.1",
     "react-markdown": "^6.0.0",
-    "react-phone-number-input": "^3.1.23",
+    "react-phone-number-input": "^3.2.19",
     "react-syntax-highlighter": "^13.0.0",
     "react-test-renderer": "18",
     "start-server-and-test": "^1.14.0",
-    "storybook-addon-performance": "^0.16.1",
+    "storybook-addon-performance": "0.16.1",
     "tar": "^4.4.8",
     "ts-jest": "^29.0.5",
-    "ts-loader": "^8.4.0",
+    "ts-loader": "^9.4.2",
+    "ts-node": "^10.9.1",
     "typescript": "4.5.5",
     "unist-builder": "^2.0.2",
     "unist-util-find": "^1.0.1",
-    "webpack": "^4.17.1",
-    "webpack-cli": "^3.1.0",
+    "webpack": "^5.75.0",
+    "webpack-cli": "^5.0.1",
     "yo": "^3.1.1",
     "yup": "^0.32.9"
   },
