@@ -1,19 +1,14 @@
 // / <reference types="Cypress" />
 
-import {routes} from '../../../site/routes';
-import {flatRoutes} from '../../support/commands';
-
-const pages = flatRoutes(routes, 'id');
-
-describe('Page accessibility', () => {
-  pages.forEach(path => {
+export function runA11yTestForRoute(route) {
+  describe(`Page accessibility for ${route} section`, () => {
     it(`should pass basic a11y test for ${Cypress.config(
       'baseUrl',
-    )}${path}`, () => {
-      cy.mockConsentAndVisit(path);
-      cy.wait(1000);
+    )}${route}`, () => {
+      cy.mockConsentAndVisit(route);
+      cy.wait(1000); // FIX: This accounts for 2mins of the test run time
       cy.injectAxe();
-      if (path === `/theme/foundation/fonts` || path === `/components/card`) {
+      if (route === `/theme/foundation/fonts` || route === `/components/card`) {
         // The typography page is a showcase so we have disabled the heading order rule for this page.
         // More info on the rule here: https://www.w3.org/WAI/tutorials/page-structure/headings/
         // TODO: Need to revisit the below color contrast rule while working on the newskit site remake project
@@ -24,9 +19,9 @@ describe('Page accessibility', () => {
           },
         });
       } else if (
-        path === `/components/tabs` ||
-        path === `/components/banner` ||
-        path === `/theme/presets/typography-presets`
+        route === `/components/tabs` ||
+        route === `/components/banner` ||
+        route === `/theme/presets/typography-presets`
       ) {
         cy.checkA11y(null, {
           rules: {
@@ -43,4 +38,4 @@ describe('Page accessibility', () => {
       }
     });
   });
-});
+}
