@@ -18,7 +18,7 @@ Object.entries(routes).forEach(([pageName, path]) => {
       cy.percySnapshot(pageName, {
         percyCSS:
           // hide the release badge as the release number is dynamic
-          'a[href^="https://github.com/newscorp-ghfb/newskit/releases/tag/"] { display: none; }',
+          'a[href^="https://github.com/newscorp-ghfb/newskit/releases/tag/"] { display: none; } #latest-release-card h1 { visibility: hidden; }',
       });
     });
   });
