@@ -35,7 +35,7 @@ const DateTimeComponent = (layoutProps: LayoutProps) => (
       introduced: 'v0.18.0',
       codeUrl:
         'https://github.com/newscorp-ghfb/newskit/tree/main/src/date-time',
-      storybookId: 'components-date-time--story-date-time',
+      storybookId: 'components-date-time--story-date-time-default',
       figmaUrl:
         'https://www.figma.com/file/FSbCQa6SzVR3K48ZWLeD77/%F0%9F%9F%A2-NK-Web-Components?node-id=2057%3A5&t=iP77Jd8O6cCJYM4p-1',
     }}
@@ -315,17 +315,25 @@ const DateTimeComponent = (layoutProps: LayoutProps) => (
               type: 'string | number',
               description: (
                 <>
-                  Timestamp string of the date to be displayed. Provide this in
-                  the ISO8601 format that is understood by the native{' '}
+                  Provide this value in a format that is understood by the
+                  native{' '}
                   <Link
                     href="https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date"
                     target="_blank"
                   >
-                    date object
+                    Date object
                   </Link>
+                  , for example, timestamp number or date string in ISO 8601
+                  format
                 </>
               ),
               required: true,
+            },
+            {
+              name: 'children',
+              type: 'string',
+              description:
+                'Provide this prop to display custom text inside the component',
             },
             {
               name: 'dateFormat',
@@ -377,7 +385,7 @@ const DateTimeComponent = (layoutProps: LayoutProps) => (
                 'If provided, this overrides the style preset applied to the prefix',
             },
             {
-              attribute: 'prefix.stypographyPreset',
+              attribute: 'prefix.typographyPreset',
               type: 'MQ<string>',
               default: 'utilityMeta020',
               description:
@@ -391,7 +399,7 @@ const DateTimeComponent = (layoutProps: LayoutProps) => (
                 'If provided, this overrides the style preset applied to the suffix',
             },
             {
-              attribute: 'suffix.stypographyPreset',
+              attribute: 'suffix.typographyPreset',
               type: 'MQ<string>',
               default: 'utilityMeta020',
               description:
