@@ -49,7 +49,7 @@ const LinkComponent = (layoutProps: LayoutProps) => (
       status: MetaStatus.Supported,
       introduced: 'v0.2.0',
       codeUrl: 'https://github.com/newscorp-ghfb/newskit/tree/main/src/link',
-      storybookId: 'components-linkinline-linkstandalone--story-link',
+      storybookId: 'components-link--story-link-default',
       figmaUrl:
         'https://www.figma.com/file/FSbCQa6SzVR3K48ZWLeD77/%F0%9F%9F%A2-NK-Web-Components?node-id=324%3A4',
     }}
