@@ -4,52 +4,30 @@ import {
   Form,
   FormInput,
   FormInputAssistiveText,
-  FormInputLabel,
-  FormInputSelect,
-  FormInputTextField,
+  FormInputCharacterCount,
   FormInputCheckbox,
+  FormInputLabel,
   FormInputRadioButton,
+  FormInputSelect,
+  FormInputTextArea,
+  FormInputTextField,
 } from '..';
 import {
   Block,
   Button,
+  FieldsetProps,
+  FormInputTextFieldProps,
+  GridLayout,
   IconButton,
-  Stack,
   styled,
+  TextBlock,
   ThemeProvider,
-  CreateThemeArgs,
-  TextFieldSize,
 } from '../..';
-import {
-  IconFilledAccountBalance,
-  IconFilledAccountTree,
-  IconFilledAddCircleOutline,
-} from '../../icons';
+import {IconFilledAddCircle} from '../../icons';
 import {SelectOption} from '../../select';
-import {
-  StorybookHeading,
-  StorybookSubHeading,
-} from '../../test/storybook-comps';
 import {Fieldset} from '../../fieldset';
 import {RadioGroup} from '../../radio-button';
 import {createCustomThemeWithBaseThemeSwitch} from '../../test/theme-select-object';
-
-const formInputCustomThemeObject: CreateThemeArgs = {
-  name: 'my-custom-select-theme',
-  overrides: {
-    stylePresets: {
-      fieldsetWithBorder: {
-        base: {
-          borderColor: '{{colors.amber010}}',
-          borderWidth: '1px',
-          borderStyle: 'solid',
-        },
-      },
-    },
-  },
-};
-
-const FormInputBlock = styled(Block)``;
 
 const sleep = (ms: number) => new Promise(resolve => setTimeout(resolve, ms));
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -64,201 +42,212 @@ const validateUserName = async (value: string) => {
   return value !== 'newskit' || 'This username is already taken';
 };
 
-export const StoryFormField = () => (
-  <>
-    <StorybookHeading>
-      FormInput with assistive text, label and before and after icon
-    </StorybookHeading>
-    <StorybookSubHeading>Valid State</StorybookSubHeading>
+const sharedProps: Pick<
+  FormInputTextFieldProps,
+  'size' | 'placeholder' | 'startEnhancer' | 'endEnhancer'
+> = {
+  size: 'small',
+  placeholder: 'Placeholder',
+  startEnhancer: (
+    <>
+      <IconFilledAddCircle overrides={{size: '20px'}} />
+    </>
+  ),
+  endEnhancer: (
+    <>
+      <IconFilledAddCircle overrides={{size: '20px'}} />
+    </>
+  ),
+};
+
+const usernameRule = {
+  required: 'Required field',
+  validate: validateUserName,
+};
+
+const emailRule = {
+  required: 'Required field',
+  pattern: {
+    // eslint-disable-next-line no-useless-escape
+    value: /^(([^<>()\[\]\.,;:\s@\"]+(\.[^<>()\[\]\.,;:\s@\"]+)*)|(\".+\"))@(([^<>()[\]\.,;:\s@\"]+\.)+[^<>()[\]\.,;:\s@\"]{2,})$/i,
+    message: 'Please provide a valid email',
+  },
+};
+
+const FormInputBlock = ({
+  label,
+  children,
+}: {
+  label?: string;
+  children: React.ReactNode;
+}) => (
+  <Block marginBlockStart="space020" marginBlockEnd="space050">
+    {label && (
+      <TextBlock
+        as="h2"
+        stylePreset="inkSubtle"
+        typographyPreset="utilityBody010"
+        marginBlockEnd="space050"
+      >
+        {label}
+      </TextBlock>
+    )}
+    {children}
+  </Block>
+);
+
+const FormInputAssistiveTextWithErrorIcon = () => (
+  <FormInputAssistiveText style={{display: 'flex', alignItems: 'center'}}>
+    <IconFilledAddCircle
+      overrides={{
+        size: '16px',
+        stylePreset: 'iconNegative',
+        marginInlineEnd: '8px',
+      }}
+    />
+    Assistive Text
+  </FormInputAssistiveText>
+);
+
+const Container = styled.div`
+  max-width: 300px;
+`;
+
+export const StoryFormInputTextField = () => (
+  <Container>
     <Form onSubmit={onSubmit}>
-      <FormInput
-        state="valid"
-        name="email-valid"
-        rules={{
-          required: 'Required field',
-          pattern: {
-            // eslint-disable-next-line no-useless-escape
-            value: /^(([^<>()\[\]\.,;:\s@\"]+(\.[^<>()\[\]\.,;:\s@\"]+)*)|(\".+\"))@(([^<>()[\]\.,;:\s@\"]+\.)+[^<>()[\]\.,;:\s@\"]{2,})$/i,
-            message: 'Please provide a valid email',
-          },
-        }}
-      >
-        <FormInputLabel>E-mail</FormInputLabel>
-        <FormInputTextField
-          size="small"
-          startEnhancer={
-            <>
-              <IconFilledAccountBalance overrides={{size: '20px'}} />
-            </>
-          }
-        />
-        <FormInputAssistiveText>Assistive Text</FormInputAssistiveText>
-      </FormInput>
-
-      <StorybookSubHeading>Invalid State</StorybookSubHeading>
-
-      <FormInput
-        state="invalid"
-        name="username-invalid"
-        rules={{
-          required: 'Required field',
-          minLength: {
-            value: 5,
-            message: 'Usernames must be at least 5 characters long',
-          },
-          validate: validateUserName,
-        }}
-      >
-        <FormInputLabel>Username</FormInputLabel>
-        <FormInputTextField
-          size="small"
-          startEnhancer={
-            <>
-              <IconFilledAccountBalance overrides={{size: 'iconSize020'}} />
-              <Block spaceInline="space020" />
-              <IconFilledAccountBalance overrides={{size: 'iconSize020'}} />
-            </>
-          }
-          endEnhancer={
-            <>
-              <IconFilledAccountBalance overrides={{size: 'iconSize020'}} />
-              <Block spaceInline="space020" />
-              <IconFilledAccountBalance overrides={{size: 'iconSize020'}} />
-            </>
-          }
-          overrides={{
-            startEnhancer: {
-              spaceInline: 'space020',
-            },
-            endEnhancer: {
-              spaceInline: 'space040',
-            },
-          }}
-        />
-        <FormInputAssistiveText>Assistive Text</FormInputAssistiveText>
-      </FormInput>
-
-      <FormInput
-        name="username-default"
-        rules={{
-          required: 'Required field',
-          minLength: {
-            value: 5,
-            message: 'Usernames must be at least 5 characters long',
-          },
-          validate: validateUserName,
-        }}
-      >
-        <StorybookSubHeading>Default State</StorybookSubHeading>
-        <FormInputLabel>Surname</FormInputLabel>
-        <FormInputTextField
-          size="small"
-          startEnhancer={
-            <IconFilledAccountBalance overrides={{size: 'iconSize020'}} />
-          }
-          endEnhancer={
-            <>
-              <IconFilledAccountBalance overrides={{size: 'iconSize020'}} />
-              <Block spaceInline="space020" />
-              <IconFilledAccountBalance overrides={{size: 'iconSize020'}} />
-            </>
-          }
-          overrides={{
-            startEnhancer: {
-              spaceInline: 'space050',
-            },
-            endEnhancer: {
-              spaceInline: 'space050',
-            },
-          }}
-        />
-        <FormInputAssistiveText>Assistive Text</FormInputAssistiveText>
-      </FormInput>
-
-      <FormInput name="password-disbled" state="disabled">
-        <StorybookSubHeading>Disabled State</StorybookSubHeading>
-        <FormInputLabel>Password</FormInputLabel>
-        <FormInputTextField
-          state="disabled"
-          size="small"
-          startEnhancer={
-            <IconFilledAccountBalance overrides={{size: '20px'}} />
-          }
-          endEnhancer={<IconFilledAccountBalance overrides={{size: '20px'}} />}
-        />
-        <FormInputAssistiveText>Assistive Text</FormInputAssistiveText>
-      </FormInput>
+      <FormInputBlock label="Base (default)">
+        <FormInput name="text-field-email" rules={emailRule}>
+          <FormInputLabel>E-mail</FormInputLabel>
+          <FormInputTextField {...sharedProps} />
+          <FormInputAssistiveText>Assistive Text</FormInputAssistiveText>
+        </FormInput>
+      </FormInputBlock>
+      <FormInputBlock label="Valid">
+        <FormInput
+          state="valid"
+          name="text-field-username"
+          rules={usernameRule}
+        >
+          <FormInputLabel>Username</FormInputLabel>
+          <FormInputTextField {...sharedProps} />
+          <FormInputAssistiveText>Assistive Text</FormInputAssistiveText>
+        </FormInput>
+      </FormInputBlock>
+      <FormInputBlock label="Invalid">
+        <FormInput state="invalid" name="text-field-first-name">
+          <FormInputLabel>First name</FormInputLabel>
+          <FormInputTextField {...sharedProps} />
+          <FormInputAssistiveTextWithErrorIcon />
+        </FormInput>
+      </FormInputBlock>
+      <FormInputBlock label="Disabled">
+        <FormInput name="text-field-surname" state="disabled">
+          <FormInputLabel>Surname</FormInputLabel>
+          <FormInputTextField {...sharedProps} />
+          <FormInputAssistiveText>Assistive Text</FormInputAssistiveText>
+        </FormInput>
+      </FormInputBlock>
       <Button type="submit">Submit</Button>
     </Form>
-  </>
+  </Container>
 );
-StoryFormField.storyName = 'form-input-before-and-after-icon';
+StoryFormInputTextField.storyName = 'Form input text field';
 
-export const FormFieldWithIconButton = () => {
-  const [showPassword, setShowPassword] = React.useState(false);
-  return (
-    <>
-      <StorybookHeading>FormInput with Button enhancer</StorybookHeading>
-      <Block>
-        <Form onSubmit={onSubmit}>
-          <FormInput
-            name="password-icon"
-            rules={{
-              required: 'Required field',
-            }}
-          >
-            <Block spaceStack="space020">
-              <FormInputLabel>Enter Password</FormInputLabel>
-            </Block>
-            <Block spaceStack="space020">
-              <FormInputTextField
-                type={!showPassword ? 'password' : 'text'}
-                startEnhancer={
-                  <IconFilledAccountBalance overrides={{size: 'iconSize020'}} />
-                }
-                endEnhancer={
-                  <IconButton
-                    aria-label="toggle password"
-                    onClick={() => setShowPassword(!showPassword)}
-                    size="small"
-                  >
-                    {showPassword ? (
-                      <IconFilledAccountBalance
-                        overrides={{size: 'iconSize010'}}
-                      />
-                    ) : (
-                      <IconFilledAccountTree
-                        overrides={{size: 'iconSize010'}}
-                      />
-                    )}
-                  </IconButton>
-                }
-                overrides={{
-                  startEnhancer: {
-                    spaceInline: 'space070',
-                  },
-                  endEnhancer: {
-                    spaceInline: 'space070',
-                  },
-                }}
-              />
-              <FormInputAssistiveText>Assistive Text</FormInputAssistiveText>
-            </Block>
-          </FormInput>
-          <Button type="submit">Submit</Button>
-        </Form>
-      </Block>
-    </>
-  );
-};
-FormFieldWithIconButton.storyName = 'form-input-with-icon-button';
+export const StoryFormInputTextArea = () => (
+  <Container>
+    <Form onSubmit={onSubmit}>
+      <FormInputBlock label="Base (default)">
+        <FormInput name="text-field-email" rules={emailRule}>
+          <FormInputLabel>E-mail</FormInputLabel>
+          <FormInputTextArea
+            {...sharedProps}
+            rows={1}
+            overrides={{height: 'unset'}}
+          />
+          <FormInputAssistiveText>Assistive Text</FormInputAssistiveText>
+        </FormInput>
+      </FormInputBlock>
+      <FormInputBlock label="Valid">
+        <FormInput
+          state="valid"
+          name="text-field-username"
+          rules={usernameRule}
+        >
+          <FormInputLabel>Username</FormInputLabel>
+          <FormInputTextArea
+            {...sharedProps}
+            rows={1}
+            overrides={{height: 'unset'}}
+          />
+          <FormInputAssistiveText>Assistive Text</FormInputAssistiveText>
+        </FormInput>
+      </FormInputBlock>
+      <FormInputBlock label="Invalid">
+        <FormInput state="invalid" name="text-field-first-name">
+          <FormInputLabel>First name</FormInputLabel>
+          <FormInputTextArea
+            {...sharedProps}
+            rows={1}
+            overrides={{height: 'unset'}}
+          />
+          <FormInputAssistiveTextWithErrorIcon />
+        </FormInput>
+      </FormInputBlock>
+      <FormInputBlock label="Disabled">
+        <FormInput name="text-field-surname" state="disabled">
+          <FormInputLabel>Surname</FormInputLabel>
+          <FormInputTextArea
+            {...sharedProps}
+            rows={1}
+            overrides={{height: 'unset'}}
+          />
+          <FormInputAssistiveText>Assistive Text</FormInputAssistiveText>
+        </FormInput>
+      </FormInputBlock>
+      <Button type="submit">Submit</Button>
+    </Form>
+  </Container>
+);
+StoryFormInputTextArea.storyName = 'Form input text area';
+
+export const FormFieldWithIconButton = () => (
+  <Container>
+    <Form onSubmit={onSubmit}>
+      <FormInputBlock>
+        <FormInput name="text-field-email" rules={emailRule}>
+          <FormInputLabel>E-mail</FormInputLabel>
+          <FormInputTextField
+            size="small"
+            placeholder="Placeholder"
+            endEnhancer={
+              <IconButton
+                aria-label="toggle password"
+                size="small"
+                overrides={{stylePreset: 'iconButtonOutlinedSecondary'}}
+              >
+                <IconFilledAddCircle
+                  overrides={{
+                    size: 'iconSize010',
+                    stylePreset: 'iconDefault',
+                  }}
+                />
+              </IconButton>
+            }
+          />
+          <FormInputAssistiveText>Assistive Text</FormInputAssistiveText>
+        </FormInput>
+      </FormInputBlock>
+      <Button type="submit">Submit</Button>
+    </Form>
+  </Container>
+);
+FormFieldWithIconButton.storyName = 'Form input with icon button';
 
 export const FormInputWithSelect = () => (
-  <>
-    <StorybookHeading>FormInput with Select</StorybookHeading>
-    <Block>
-      <StorybookSubHeading>Standalone Select</StorybookSubHeading>
-      <Form onSubmit={onSubmit}>
+  <Container>
+    <Form onSubmit={onSubmit}>
+      <FormInputBlock label="Standalone select">
         <FormInput
           name="select"
           rules={{
@@ -269,19 +258,15 @@ export const FormInputWithSelect = () => (
             <FormInputLabel>Select Item</FormInputLabel>
           </Block>
           <Block spaceStack="space020">
-            <FormInputSelect overrides={{button: {width: '300px'}}}>
+            <FormInputSelect size="small" placeholder="Select option">
               <SelectOption value="Option 1">Option 1</SelectOption>
               <SelectOption value="Option 2">Option 2</SelectOption>
             </FormInputSelect>
             <FormInputAssistiveText>Assistive Text</FormInputAssistiveText>
           </Block>
         </FormInput>
-        <Button type="submit">Submit</Button>
-      </Form>
-    </Block>
-    <Block>
-      <StorybookSubHeading>Select with enhancers</StorybookSubHeading>
-      <Form onSubmit={onSubmit}>
+      </FormInputBlock>
+      <FormInputBlock label="Start and end enhancers">
         <FormInput
           name="select"
           rules={{
@@ -293,13 +278,9 @@ export const FormInputWithSelect = () => (
           </Block>
           <Block spaceStack="space020">
             <FormInputSelect
-              overrides={{button: {width: '300px'}}}
-              startEnhancer={
-                <IconFilledAccountBalance overrides={{size: 'iconSize020'}} />
-              }
-              endEnhancer={
-                <IconFilledAccountBalance overrides={{size: 'iconSize020'}} />
-              }
+              {...sharedProps}
+              placeholder="Select option"
+              endEnhancer={undefined}
             >
               <SelectOption value="Option 1">Option 1</SelectOption>
               <SelectOption value="Option 2">Option 2</SelectOption>
@@ -307,509 +288,195 @@ export const FormInputWithSelect = () => (
             <FormInputAssistiveText>Assistive Text</FormInputAssistiveText>
           </Block>
         </FormInput>
-        <Button type="submit">Submit</Button>
-      </Form>
-    </Block>
-  </>
+      </FormInputBlock>
+      <Button type="submit">Submit</Button>
+    </Form>
+  </Container>
 );
-FormInputWithSelect.storyName = 'form-input-with-select';
-
-const spaceStack = {
-  small: 'space020',
-  medium: 'space020',
-  large: 'space030',
-};
+FormInputWithSelect.storyName = 'Form input select';
 
 export const FormFieldCheckbox = () => (
   <Form onSubmit={onSubmit}>
-    <StorybookSubHeading>Single checkbox validation</StorybookSubHeading>
-    <Block paddingBlock="space050" paddingInline="space050">
-      {['small', 'medium', 'large'].map(size => (
-        <Block spaceStack="space050">
-          <FormInput
-            name={`${size}-name`}
-            size={size as TextFieldSize}
-            rules={{
-              required: 'Required field',
-            }}
-          >
-            <FormInputCheckbox
-              label={`Agree with T&C ${size}`}
-              value="tc"
-              overrides={{
-                spaceStack: spaceStack[size as keyof typeof spaceStack],
-              }}
-            />
-            <FormInputAssistiveText validationIcon>
-              Please agree with our T&C
-            </FormInputAssistiveText>
-          </FormInput>
-        </Block>
-      ))}
-
-      <Button type="submit">Submit</Button>
-    </Block>
+    <FormInputBlock>
+      <FormInput
+        name="checkbox"
+        size="small"
+        rules={{
+          required: 'Required field',
+        }}
+      >
+        <FormInputCheckbox
+          label="I agree to the terms & conditions"
+          overrides={{marginBlockEnd: '7px'}}
+        />
+        <FormInputAssistiveText validationIcon>
+          Please agree with our terms & conditions
+        </FormInputAssistiveText>
+      </FormInput>
+    </FormInputBlock>
+    <Button type="submit">Submit</Button>
   </Form>
 );
-FormFieldCheckbox.storyName = 'form-input-checkbox';
+FormFieldCheckbox.storyName = 'Form input checkbox';
 
-export const FormFieldRadioButton = () => {
-  const rules = {
-    required: 'Required field',
-  };
-
-  return (
-    <Block paddingBlock="space050" paddingInline="space050">
-      <Form onSubmit={onSubmit}>
-        <StorybookHeading>RadioButton and RadioGroup</StorybookHeading>
-        <Fieldset legend="Favorite activities">
-          <RadioGroup>
-            {['Reading', 'Writing', 'Speaking', 'Listening'].map(value => (
-              <FormInput key={value} name="favoriteActivities" rules={rules}>
-                <FormInputRadioButton
-                  label={value}
-                  value={value}
-                  overrides={{spaceStack: 'space030'}}
-                />
-              </FormInput>
-            ))}
-          </RadioGroup>
-          <FormInput name="favoriteActivities" rules={rules}>
-            <FormInputAssistiveText>Assistive Text</FormInputAssistiveText>
+export const FormFieldRadioButton = () => (
+  <Form onSubmit={onSubmit}>
+    <FormInputBlock>
+      <Fieldset legend="Select an option">
+        <RadioGroup>
+          <FormInput name="option1" rules={{required: 'Required field'}}>
+            <FormInputRadioButton
+              label="Option 1"
+              value="1"
+              overrides={{spaceStack: 'space030'}}
+            />
           </FormInput>
-        </Fieldset>
-        <Block spaceStack="space030" />
-        <Button type="submit">Submit</Button>
-      </Form>
-    </Block>
+          <FormInput name="option2" rules={{required: 'Required field'}}>
+            <FormInputRadioButton
+              label="Option 2"
+              value="2"
+              overrides={{spaceStack: 'space030'}}
+            />
+          </FormInput>
+          <FormInput name="option3" rules={{required: 'Required field'}}>
+            <FormInputRadioButton
+              label="Option 3"
+              value="3"
+              overrides={{spaceStack: 'space030'}}
+            />
+          </FormInput>
+        </RadioGroup>
+        <FormInput
+          name="favoriteActivities"
+          rules={{required: 'Required field'}}
+        >
+          <FormInputAssistiveText>Assistive Text</FormInputAssistiveText>
+        </FormInput>
+      </Fieldset>
+    </FormInputBlock>
+    <Button type="submit">Submit</Button>
+  </Form>
+);
+
+FormFieldRadioButton.storyName = 'Form input radio button';
+
+export const StoryFormFieldset = () => {
+  const sizes: NonNullable<FieldsetProps['size']>[] = [
+    'small',
+    'medium',
+    'large',
+  ];
+  return (
+    <GridLayout columns={{sm: '1fr', md: '1fr 1fr', lg: '1fr 1fr 1fr'}}>
+      {sizes.map(size => (
+        <Container>
+          <Form onSubmit={onSubmit} marginBlockEnd="space050">
+            <Fieldset
+              legend={`${size.slice(0, 1).toUpperCase()}${size.substring(1)}`}
+              size={size}
+            >
+              <FormInputBlock>
+                <FormInput name="first-name" rules={emailRule}>
+                  <FormInputLabel>Email</FormInputLabel>
+                  <FormInputTextField size={size} placeholder="Placeholder" />
+                  <FormInputAssistiveText>
+                    Assistive text
+                  </FormInputAssistiveText>
+                </FormInput>
+              </FormInputBlock>
+              <FormInputBlock>
+                <FormInput name="last-name" rules={usernameRule}>
+                  <FormInputLabel>Username</FormInputLabel>
+                  <FormInputTextField size={size} placeholder="Placeholder" />
+                  <FormInputAssistiveText>
+                    Assistive text
+                  </FormInputAssistiveText>
+                </FormInput>
+              </FormInputBlock>
+              <Button type="submit" size={size}>
+                Submit
+              </Button>
+            </Fieldset>
+          </Form>
+        </Container>
+      ))}
+    </GridLayout>
   );
 };
+StoryFormFieldset.storyName = 'Form input fieldset';
 
-FormFieldRadioButton.storyName = 'form-input-radio-button';
-
-export const StoryFormFieldset = () => (
-  <>
-    <StorybookHeading>FormInputs inside Fieldset</StorybookHeading>
-
+export const StoryFormInputCharacterCount = () => (
+  <Container>
     <Form onSubmit={onSubmit}>
-      <Stack stackDistribution="space-between" flow="horizontal-top">
-        <Fieldset
-          legend="Small Fieldset"
-          size="small"
-          overrides={{
-            stylePreset: 'fieldsetWithBorder',
-            legend: {
-              spaceStack: 'space070',
+      <FormInputBlock label="Base (default)">
+        <FormInput
+          name="text-field-email"
+          rules={{
+            ...emailRule,
+            maxLength: {
+              value: 200,
             },
           }}
         >
-          <FormInputBlock spaceStack="space040">
-            <FormInput
-              name="first-name"
-              rules={{
-                required: 'Required field',
-                minLength: {
-                  value: 2,
-                  message: 'Name must be at least 2 characters long',
-                },
-                validate: validateUserName,
-              }}
-            >
-              <FormInputLabel
-                overrides={{
-                  typographyPreset: 'utilityLabel010',
-                  spaceStack: 'space030',
-                }}
-              >
-                First name
-              </FormInputLabel>
-              <FormInputTextField
-                startEnhancer={
-                  <IconFilledAddCircleOutline
-                    overrides={{size: 'iconSize020'}}
-                  />
-                }
-                overrides={{spaceStack: 'space000'}}
-              />
-            </FormInput>
-          </FormInputBlock>
-          <FormInputBlock spaceStack="space050">
-            <FormInput
-              name="last-name"
-              rules={{
-                required: 'Required field',
-                minLength: {
-                  value: 2,
-                  message: 'Last name must be at least 2 characters long',
-                },
-                validate: validateUserName,
-              }}
-            >
-              <FormInputLabel
-                overrides={{
-                  typographyPreset: 'utilityLabel010',
-                  spaceStack: 'space030',
-                }}
-              >
-                Last name
-              </FormInputLabel>
-              <FormInputTextField
-                startEnhancer={
-                  <IconFilledAddCircleOutline
-                    overrides={{size: 'iconSize020'}}
-                  />
-                }
-                overrides={{spaceStack: 'space020'}}
-              />
-              <FormInputAssistiveText>
-                This should be your legal last name
-              </FormInputAssistiveText>
-            </FormInput>
-          </FormInputBlock>
-          <FormInputBlock>
-            <FormInput
-              name="small-checkbox"
-              rules={{
-                required: 'Required field',
-              }}
-            >
-              <FormInputCheckbox
-                label="I agree to the terms & conditions"
-                value="tc"
-                overrides={{spaceStack: 'space020'}}
-              />
-            </FormInput>
-            <FormInput
-              name="small-checkbox"
-              rules={{
-                required: 'Required field',
-              }}
-            >
-              <FormInputCheckbox
-                label="I want to hear about more offers"
-                value="tc"
-              />
-            </FormInput>
-          </FormInputBlock>
-        </Fieldset>
-
-        <Fieldset
-          legend="Medium Fieldset"
-          size="medium"
-          overrides={{
-            stylePreset: 'fieldsetWithBorder',
-            legend: {
-              spaceStack: 'space070',
+          <FormInputLabel>E-mail</FormInputLabel>
+          <FormInputTextField {...sharedProps} />
+          <FormInputCharacterCount />
+        </FormInput>
+      </FormInputBlock>
+      <FormInputBlock label="Valid">
+        <FormInput
+          state="valid"
+          name="text-field-username"
+          rules={{
+            ...usernameRule,
+            maxLength: {
+              value: 40,
             },
           }}
         >
-          <FormInputBlock spaceStack="space050">
-            <FormInput
-              name="first-name"
-              rules={{
-                required: 'Required field',
-                minLength: {
-                  value: 2,
-                  message: 'Name must be at least 2 characters long',
-                },
-                validate: validateUserName,
-              }}
-            >
-              <FormInputLabel
-                overrides={{
-                  typographyPreset: 'utilityLabel020',
-                  spaceStack: 'space030',
-                }}
-              >
-                First name
-              </FormInputLabel>
-              <FormInputTextField
-                startEnhancer={
-                  <IconFilledAddCircleOutline
-                    overrides={{size: 'iconSize020'}}
-                  />
-                }
-                overrides={{spaceStack: 'space000'}}
-              />
-            </FormInput>
-          </FormInputBlock>
-
-          <FormInputBlock spaceStack="space060">
-            <FormInput
-              name="last-name"
-              rules={{
-                required: 'Required field',
-                minLength: {
-                  value: 2,
-                  message: 'Last name must be at least 2 characters long',
-                },
-                validate: validateUserName,
-              }}
-            >
-              <FormInputLabel
-                overrides={{
-                  typographyPreset: 'utilityLabel020',
-                  spaceStack: 'space030',
-                }}
-              >
-                Last name
-              </FormInputLabel>
-              <FormInputTextField
-                startEnhancer={
-                  <IconFilledAddCircleOutline
-                    overrides={{size: 'iconSize020'}}
-                  />
-                }
-                overrides={{spaceStack: 'space020'}}
-              />
-              <FormInputAssistiveText>
-                This should be your legal last name
-              </FormInputAssistiveText>
-            </FormInput>
-          </FormInputBlock>
-
-          <FormInputBlock>
-            <FormInput
-              name="medium-checkbox"
-              rules={{
-                required: 'Required field',
-              }}
-            >
-              <FormInputCheckbox
-                label="I agree to the terms & conditions"
-                value="tc"
-                overrides={{spaceStack: 'space030'}}
-              />
-            </FormInput>
-            <FormInput
-              name="medium-checkbox"
-              rules={{
-                required: 'Required field',
-              }}
-            >
-              <FormInputCheckbox
-                label="I want to hear about more offers"
-                value="tc"
-              />
-            </FormInput>
-          </FormInputBlock>
-        </Fieldset>
-
-        <Fieldset
-          legend="Large Fieldset"
-          size="large"
-          overrides={{
-            stylePreset: 'fieldsetWithBorder',
-            legend: {
-              spaceStack: 'space070',
+          <FormInputLabel>Username</FormInputLabel>
+          <FormInputTextField {...sharedProps} />
+          <FormInputCharacterCount />
+        </FormInput>
+      </FormInputBlock>
+      <FormInputBlock label="Invalid">
+        <FormInput
+          state="invalid"
+          name="text-field-first-name"
+          rules={{
+            maxLength: {
+              value: -15,
             },
           }}
         >
-          <FormInputBlock spaceStack="space050">
-            <FormInput
-              name="first-name"
-              rules={{
-                required: 'Required field',
-                minLength: {
-                  value: 2,
-                  message: 'Name must be at least 2 characters long',
-                },
-                validate: validateUserName,
-              }}
-            >
-              <FormInputLabel
-                overrides={{
-                  typographyPreset: 'utilityLabel030',
-                  spaceStack: 'space030',
-                }}
-              >
-                First name
-              </FormInputLabel>
-              <FormInputTextField
-                startEnhancer={
-                  <IconFilledAddCircleOutline
-                    overrides={{size: 'iconSize010'}}
-                  />
-                }
-                overrides={{
-                  spaceStack: 'space000',
-                }}
-              />
-            </FormInput>
-          </FormInputBlock>
-          <FormInputBlock spaceStack="space060">
-            <FormInput
-              name="last-name"
-              rules={{
-                required: 'Required field',
-                minLength: {
-                  value: 2,
-                  message: 'Last name must be at least 2 characters long',
-                },
-                validate: validateUserName,
-              }}
-            >
-              <FormInputLabel
-                overrides={{
-                  typographyPreset: 'utilityLabel030',
-                  spaceStack: 'space030',
-                }}
-              >
-                Last name
-              </FormInputLabel>
-              <FormInputTextField
-                startEnhancer={
-                  <IconFilledAddCircleOutline
-                    overrides={{size: 'iconSize010'}}
-                  />
-                }
-                overrides={{
-                  spaceStack: 'space020',
-                }}
-              />
-              <FormInputAssistiveText>
-                This should be your legal last name
-              </FormInputAssistiveText>
-            </FormInput>
-          </FormInputBlock>
-
-          <FormInputBlock>
-            <FormInput
-              name="large-checkbox"
-              rules={{
-                required: 'Required field',
-              }}
-            >
-              <FormInputCheckbox
-                label="I agree to the terms & conditions"
-                value="tc"
-                overrides={{spaceStack: 'space030'}}
-              />
-            </FormInput>
-            <FormInput
-              name="large-checkbox"
-              rules={{
-                required: 'Required field',
-              }}
-            >
-              <FormInputCheckbox
-                label="I want to hear about more offers"
-                value="tc"
-              />
-            </FormInput>
-          </FormInputBlock>
-        </Fieldset>
-
-        <Fieldset
-          legend="Mixed Size Fieldset"
-          size="small"
-          overrides={{
-            stylePreset: 'fieldsetWithBorder',
-            legend: {
-              spaceStack: 'space070',
+          <FormInputLabel>First name</FormInputLabel>
+          <FormInputTextField {...sharedProps} />
+          <FormInputCharacterCount />
+        </FormInput>
+      </FormInputBlock>
+      <FormInputBlock label="Disabled">
+        <FormInput
+          name="text-field-surname"
+          state="disabled"
+          rules={{
+            maxLength: {
+              value: 200,
             },
           }}
         >
-          <FormInputBlock spaceStack="space040">
-            <FormInput
-              name="first-name"
-              rules={{
-                required: 'Required field',
-                minLength: {
-                  value: 2,
-                  message: 'Name must be at least 2 characters long',
-                },
-                validate: validateUserName,
-              }}
-            >
-              <FormInputLabel
-                overrides={{
-                  typographyPreset: 'utilityLabel010',
-                  spaceStack: 'space030',
-                }}
-              >
-                First name
-              </FormInputLabel>
-              <FormInputTextField
-                startEnhancer={
-                  <IconFilledAddCircleOutline
-                    overrides={{size: 'iconSize020'}}
-                  />
-                }
-                overrides={{spaceStack: 'space000'}}
-              />
-            </FormInput>
-          </FormInputBlock>
-          <FormInputBlock spaceStack="space050">
-            <FormInput
-              name="last-name"
-              rules={{
-                required: 'Required field',
-                minLength: {
-                  value: 2,
-                  message: 'Last name must be at least 2 characters long',
-                },
-                validate: validateUserName,
-              }}
-              size="large"
-            >
-              <FormInputLabel
-                overrides={{
-                  typographyPreset: 'utilityLabel010',
-                  spaceStack: 'space030',
-                }}
-              >
-                Last name
-              </FormInputLabel>
-              <FormInputTextField
-                startEnhancer={
-                  <IconFilledAddCircleOutline
-                    overrides={{size: 'iconSize020'}}
-                  />
-                }
-                overrides={{spaceStack: 'space020'}}
-              />
-              <FormInputAssistiveText>
-                This should be your legal last name
-              </FormInputAssistiveText>
-            </FormInput>
-          </FormInputBlock>
-          <FormInputBlock>
-            <FormInput
-              name="medium-checkbox"
-              rules={{
-                required: 'Required field',
-              }}
-              size="large"
-            >
-              <FormInputCheckbox
-                label="I agree to the terms & conditions"
-                value="tc"
-                overrides={{spaceStack: 'space020'}}
-                size="medium"
-              />
-            </FormInput>
-            <FormInput
-              name="small-checkbox"
-              rules={{
-                required: 'Required field',
-              }}
-            >
-              <FormInputCheckbox
-                label="I want to hear about more offers"
-                value="tc"
-              />
-            </FormInput>
-          </FormInputBlock>
-        </Fieldset>
-      </Stack>
+          <FormInputLabel>Surname</FormInputLabel>
+          <FormInputTextField {...sharedProps} />
+          <FormInputCharacterCount aria-disabled />
+        </FormInput>
+      </FormInputBlock>
+      <Button type="submit">Submit</Button>
     </Form>
-  </>
+  </Container>
 );
-StoryFormFieldset.storyName = 'form-input-fieldset';
+StoryFormInputCharacterCount.storyName = 'Form input character count';
 
 export default {
-  title: 'Components/form-input',
+  title: 'Components/Form input',
   component: () => 'None',
   disabledRules: [],
   decorators: [
@@ -817,11 +484,19 @@ export default {
       <ThemeProvider
         theme={createCustomThemeWithBaseThemeSwitch(
           context?.globals?.backgrounds?.value,
-          formInputCustomThemeObject,
+          {},
         )}
       >
         <Story />
       </ThemeProvider>
     ),
   ],
+  parameters: {
+    nkDocs: {
+      title: 'Form input',
+      url: 'https://newskit.co.uk/components/form',
+      description:
+        'The form component allows users to enter and edit information into a UI using form controls. Based on React Hook Form.',
+    },
+  },
 };
