@@ -2,49 +2,29 @@
 import * as React from 'react';
 import {Story as StoryType} from '@storybook/react';
 
-import {
-  AccountBalance as OutlinedAccountBalance,
-  AccountTree as OutlinedAccountTree,
-} from '@emotion-icons/material-outlined';
-import {
-  AccountBalance as FilledAccountBalance,
-  AccountTree as FilledAccountTree,
-} from '@emotion-icons/material';
-
 import {createTheme, ThemeProvider, CreateThemeArgs} from '../../theme';
-import {styled} from '../../utils/style';
-import {Stack} from '../../stack';
-import {StorybookSubHeading} from '../../test/storybook-comps';
+import {styled, getColorCssFromTheme} from '../../utils/style';
+
+import {StorybookCase, StorybookPage} from '../../test/storybook-comps';
 import {createCustomThemeWithBaseThemeSwitch} from '../../test/theme-select-object';
 
-import {IconOutlinedClose, Svg, NewsKitIcon} from '..';
+import {
+  IconOutlinedClose,
+  Svg,
+  NewsKitIcon,
+  IconFilledAddCircle,
+  IconFilledAddCircleOutline,
+} from '..';
 import * as customIcons from '../filled/custom';
 import {customToNewsKitIcon} from '../custom-to-newskit-icon';
+import {InlineMessage} from '../../inline-message';
+import {GridLayout, GridLayoutItem} from '../../grid-layout';
+import {LinkInline} from '../../link';
 
-const IconFilledAccountBalance = customToNewsKitIcon(
-  'IconFilledAccountBalance',
-  props => <FilledAccountBalance {...props} />,
-);
-const IconOutlinedAccountBalance = customToNewsKitIcon(
-  'IconOutlinedAccountBalance',
-  props => <OutlinedAccountBalance {...props} />,
-);
-const IconFilledAccountTree = customToNewsKitIcon(
-  'IconFilledAccountTree',
-  props => <FilledAccountTree {...props} />,
-);
-const IconOutlinedAccountTree = customToNewsKitIcon(
-  'IconOutlinedAccountTree',
-  props => <OutlinedAccountTree {...props} />,
-);
-
-const Container = styled.div<{dark?: boolean}>`
-  ${({dark}) => dark && 'background: #000;'}
-`;
-
-const Constrain = styled.div`
-  width: 60px;
-  display: inline-block;
+const MarginOverridesWrapper = styled.div`
+  width: min-content;
+  border: 1px dashed;
+  ${getColorCssFromTheme('borderColor', 'red060')}
 `;
 
 const iconsCustomThemeObject: CreateThemeArgs = {
@@ -53,185 +33,155 @@ const iconsCustomThemeObject: CreateThemeArgs = {
     stylePresets: {
       iconCustom: {
         base: {
-          iconColor: '{{colors.amber050}}',
+          iconColor: '{{colors.inkBase}}',
         },
       },
       iconCustomWithTransitions: {
         base: {
-          iconColor: '{{colors.amber050}}',
+          iconColor: '{{colors.inkBase}}',
         },
         hover: {
-          iconColor: '{{colors.green050}}',
+          iconColor: '{{colors.interactivePrimary030}}',
         },
       },
     },
   },
 };
 
-const getSocialIconStylePreset = (iconName: string) =>
-  `iconSocial${iconName
-    .split(/(?=[A-Z])/)
-    .filter(x => !/Icon|Filled|Outlined /.test(x))
-    .join('')}`;
-
 // IndeterminateProgressIndicator is excluded from Storybook because visual testing of animated
 // components like ProgressIndicator (https://newskit.co.uk/components/progress-indicator/)
 // is always likely to be problematic
 const {IndeterminateProgressIndicator, ...remainingCustomIcons} = customIcons;
 
-const materialIconsSample = {
-  IconFilledAccountBalance,
-  IconOutlinedAccountBalance,
-  IconFilledAccountTree,
-  IconOutlinedAccountTree,
-};
+export const StoryDefaultIcons = () => (
+  <StorybookPage>
+    <StorybookCase>
+      <IconFilledAddCircle
+        overrides={{
+          size: 'iconSize040',
+        }}
+      />
+    </StorybookCase>
+  </StorybookPage>
+);
+StoryDefaultIcons.storyName = 'Default';
 
-const socialIconList = [
-  'IconFilledFacebook',
-  'IconFilledGitHub',
-  'IconFilledTwitter',
-  'IconFilledWhatsApp',
-  'IconFilledReddit',
-  'IconFilledInstagram',
-  'IconFilledYoutube',
-];
-
-const isSocialIcon = (iconName: string) => socialIconList.includes(iconName);
+export const StorySizeIcons = () => (
+  <StorybookPage
+    columns="repeat(5, 1fr)"
+    columnGap="space080"
+    rowGap="space080"
+  >
+    <StorybookCase title="iconSize005 (8px)">
+      <IconFilledAddCircle
+        overrides={{
+          size: 'iconSize005',
+        }}
+      />
+    </StorybookCase>
+    <StorybookCase title="iconSize010 (16px)">
+      <IconFilledAddCircle
+        overrides={{
+          size: 'iconSize010',
+        }}
+      />
+    </StorybookCase>
+    <StorybookCase title="iconSize020 (24px)">
+      <IconFilledAddCircle
+        overrides={{
+          size: 'iconSize020',
+        }}
+      />
+    </StorybookCase>
+    <StorybookCase title="iconSize030 (32px)">
+      <IconFilledAddCircle
+        overrides={{
+          size: 'iconSize030',
+        }}
+      />
+    </StorybookCase>
+    <StorybookCase title="iconSize040 (48px)">
+      <IconFilledAddCircle
+        overrides={{
+          size: 'iconSize040',
+        }}
+      />
+    </StorybookCase>
+    <StorybookCase title="iconSize050 (64px)">
+      <IconFilledAddCircle
+        overrides={{
+          size: 'iconSize050',
+        }}
+      />
+    </StorybookCase>
+  </StorybookPage>
+);
+StorySizeIcons.storyName = 'Size';
 
 export const StoryCustomIcons = () => (
-  <>
+  <StorybookPage columns="repeat(5, 1fr)">
     {Object.entries(remainingCustomIcons).map((entry: any) => {
       const [iconName, Icon] = entry as [string, NewsKitIcon];
 
-      const stylePresetOverride = isSocialIcon(iconName)
-        ? getSocialIconStylePreset(iconName)
-        : 'iconCustom';
-
       return (
-        <Stack
-          flow="horizontal-bottom"
-          wrap="wrap"
-          spaceInline="sizing020"
-          key={iconName}
-        >
-          <Container>
-            <Constrain>
-              <Icon />
-            </Constrain>
-            <Icon overrides={{size: 'iconSize020'}} />
-            <Icon
-              overrides={{
-                size: {
-                  xs: 'iconSize010',
-                  sm: 'iconSize020',
-                  md: 'iconSize030',
-                  lg: 'iconSize040',
-                  xl: 'iconSize050',
-                },
-              }}
-            />
-          </Container>
-          <Container>
-            <Constrain>
-              <Icon
-                overrides={{
-                  stylePreset: stylePresetOverride,
-                }}
-              />
-            </Constrain>
-            <Icon
-              overrides={{
-                size: 'iconSize020',
-                stylePreset: stylePresetOverride,
-              }}
-            />
-            {/* </ThemeProvider> */}
-          </Container>
-          <Container dark>
-            {/* <ThemeProvider theme={getTheme(newskitDarkTheme)}> */}
-            <Constrain>
-              <Icon
-                overrides={{
-                  stylePreset: stylePresetOverride,
-                }}
-              />
-            </Constrain>
-            <Icon
-              overrides={{
-                size: 'iconSize020',
-                stylePreset: stylePresetOverride,
-              }}
-            />
-          </Container>
-        </Stack>
+        <StorybookCase title={iconName}>
+          <Icon overrides={{size: 'iconSize040'}} />
+        </StorybookCase>
       );
     })}
-  </>
+  </StorybookPage>
 );
-StoryCustomIcons.storyName = 'custom-icons';
+StoryCustomIcons.storyName = 'Custom icons';
 
 export const StoryMaterialIcons = () => (
-  <>
-    {Object.values(materialIconsSample).map((Icon: NewsKitIcon) => (
-      <div>
-        <Constrain>
-          <Icon />
-        </Constrain>
-        <Icon overrides={{size: 'iconSize030'}} />
-        <Icon
-          overrides={{
-            size: {
-              xs: 'iconSize010',
-              sm: 'iconSize020',
-              md: 'iconSize030',
-              lg: 'iconSize040',
-              xl: 'iconSize050',
-            },
-          }}
-        />
-
-        <Constrain>
-          <Icon
-            overrides={{
-              stylePreset: 'iconCustom',
-            }}
-          />
-        </Constrain>
-        <Icon
-          overrides={{
-            size: 'iconSize020',
-            stylePreset: 'iconCustom',
-          }}
-        />
-      </div>
-    ))}
-  </>
+  <GridLayout columns="1fr 1fr" columnGap="space080" rowGap="space080">
+    <StorybookCase title="IconFilledAddCircle">
+      <IconFilledAddCircle overrides={{size: 'iconSize040'}} />
+    </StorybookCase>
+    <StorybookCase title="IconFilledAddCircleOutline">
+      <IconFilledAddCircleOutline overrides={{size: 'iconSize040'}} />
+    </StorybookCase>
+    <GridLayoutItem column="1 / 3">
+      <InlineMessage>
+        For the full library of Material icons available, please refer to{' '}
+        <LinkInline href="fonts.google.com/icons">
+          fonts.google.com/icons
+        </LinkInline>
+      </InlineMessage>
+    </GridLayoutItem>
+  </GridLayout>
 );
-StoryMaterialIcons.storyName = 'material-icons';
+StoryMaterialIcons.storyName = 'Material icons';
+
+export const StoryIconsTransition = () => (
+  <StorybookPage>
+    <StorybookCase title="Transition preset override on hover">
+      <IconFilledAddCircle
+        overrides={{
+          size: 'iconSize040',
+          stylePreset: 'iconCustomWithTransitions',
+        }}
+      />
+    </StorybookCase>
+  </StorybookPage>
+);
+
+StoryIconsTransition.storyName = 'Transitions';
 
 export const StoryOverriddenIcons = () => {
   const MyCustomIcon = customToNewsKitIcon('IconFilledStop', props => (
-    <Svg {...props} viewBox="0 0 429 422">
-      <g
-        fill="none"
-        fillRule="evenodd"
-        transform="translate(-77.000000, -1738.000000) translate(77.000000, 1738.500000)"
-      >
-        <circle
-          fill="#F0A4A4"
-          cx="250.5"
-          cy="242.5"
-          r="178"
-          style={{mixBlendMode: 'multiply'}}
-        />
-        <circle
-          fill="#5166CE"
-          cx="195"
-          cy="195"
-          r="195"
-          style={{mixBlendMode: 'multiply'}}
-        />
-      </g>
+    <Svg {...props} viewBox="0 0 48 48">
+      <circle opacity="0.3" cx="27.5" cy="27.5" r="19.5" fill="#D60000" />
+      <circle opacity="0.3" cx="21.5" cy="20.5" r="19.5" fill="#007B22" />
+      <path
+        opacity="0.8"
+        d="M24.5 4.5C13.7478 4.5 5 13.2478 5 24C5 34.7522 13.7478 43.5 24.5 43.5C35.2522 43.5 44 34.7522 44 24C44 13.2478 35.2522 4.5 24.5 4.5ZM32.6216 30L30.5 32.1216L24.5 26.1216L18.5 32.1216L16.3784 30L22.3784 24L16.3784 18L18.5 15.8784L24.5 21.8784L30.5 15.8784L32.6216 18L26.6216 24L32.6216 30Z"
+        fill="#3358CC"
+      />
+      <path
+        d="M32.622 30L30.5005 32.1215L24.5005 26.1215L18.5005 32.1215L16.3789 30L22.3789 24L16.3789 18L18.5005 15.8784L24.5005 21.8784L30.5005 15.8784L32.622 18L26.622 24L32.622 30Z"
+        fill="white"
+      />
     </Svg>
   ));
   const themeWithOverridenIcons = createTheme({
@@ -245,59 +195,50 @@ export const StoryOverriddenIcons = () => {
   });
 
   return (
-    <>
-      <StorybookSubHeading>IconOutlinedClose default</StorybookSubHeading>
-      <IconOutlinedClose
-        overrides={{
-          size: 'iconSize050',
-          stylePreset: 'iconSocialFacebook',
-        }}
-      />
-      <StorybookSubHeading>IconOutlinedClose overridden</StorybookSubHeading>
-      <ThemeProvider theme={themeWithOverridenIcons}>
+    <StorybookPage>
+      <StorybookCase title="IconOutlinedClose default">
         <IconOutlinedClose
           overrides={{
             size: 'iconSize050',
-            stylePreset: 'iconSocialFacebook',
           }}
         />
-      </ThemeProvider>
-    </>
+      </StorybookCase>
+      <StorybookCase title="IconOutlinedClose overriden">
+        <ThemeProvider theme={themeWithOverridenIcons}>
+          <IconOutlinedClose
+            overrides={{
+              size: 'iconSize050',
+            }}
+          />
+        </ThemeProvider>
+      </StorybookCase>
+      <StorybookCase title="Logical props">
+        <MarginOverridesWrapper>
+          <IconFilledAddCircle
+            overrides={{
+              marginBlock: 'space010',
+              marginInline: 'space010',
+              size: 'iconSize050',
+            }}
+          />
+        </MarginOverridesWrapper>
+      </StorybookCase>
+    </StorybookPage>
   );
 };
-StoryOverriddenIcons.storyName = 'overridden-icons';
-
-export const StoryIconsTransition = () => (
-  <>
-    {Object.values([IconFilledAccountBalance]).map((Icon: NewsKitIcon) => (
-      <Constrain>
-        <StorybookSubHeading>Icon with transition</StorybookSubHeading>
-        <Icon
-          overrides={{
-            stylePreset: 'iconCustomWithTransitions',
-          }}
-        />
-      </Constrain>
-    ))}
-  </>
-);
-StoryIconsTransition.storyName = 'icons-transition-overrides';
-
-export const StoryIconsLogicalProps = () => (
-  <Constrain>
-    <StorybookSubHeading>Icon with logical props</StorybookSubHeading>
-    <IconFilledAccountBalance
-      overrides={{
-        marginBlock: '20px',
-      }}
-    />
-  </Constrain>
-);
-StoryIconsLogicalProps.storyName = 'icons-logical-props';
+StoryOverriddenIcons.storyName = 'Overrides';
 
 export default {
-  title: 'Components/icons',
+  title: 'Components/Icons',
   component: () => 'None',
+  parameters: {
+    nkDocs: {
+      title: 'Icons',
+      url: 'https://www.newskit.co.uk/components/icons/',
+      description:
+        'Icons are small SVG shapes, ranging from basic UI shapes to brand logos.',
+    },
+  },
   decorators: [
     (Story: StoryType, context: {globals: {backgrounds: {value: string}}}) => (
       <ThemeProvider
