@@ -1,26 +1,24 @@
 // Social custom icons
-export * from './icon-filled-figma';
 export * from './icon-filled-facebook';
-export * from './icon-filled-github';
-export * from './icon-filled-instagram';
-export * from './icon-filled-reddit';
-export * from './icon-filled-twitter';
-export * from './icon-filled-whatsapp';
-export * from './icon-filled-youtube';
-
+export * from './icon-filled-figma';
 export * from './icon-filled-forward-5';
 export * from './icon-filled-forward-10';
 export * from './icon-filled-forward-15';
 export * from './icon-filled-forward-20';
 export * from './icon-filled-forward-25';
 export * from './icon-filled-forward-30';
-
+export * from './icon-filled-github';
+export * from './icon-filled-instagram';
+export * from './icon-filled-reddit';
 export * from './icon-filled-replay-5';
 export * from './icon-filled-replay-10';
 export * from './icon-filled-replay-15';
 export * from './icon-filled-replay-20';
 export * from './icon-filled-replay-25';
 export * from './icon-filled-replay-30';
+export * from './icon-filled-twitter';
+export * from './icon-filled-whatsapp';
+export * from './icon-filled-youtube';
 
 // Other custom icons
 export * from './indeterminate-progress-indicator';
