@@ -1,5 +1,6 @@
-import React from 'react';
+import React, {useRef} from 'react';
 import {Story as StoryType} from '@storybook/react';
+import {useTheme, CreateThemeArgs, ThemeProvider} from '../../theme';
 import {Pagination} from '../pagination';
 import {PaginationFirstItem} from '../components/first-item/first-item';
 import {PaginationPrevItem} from '../components/prev-item/prev-item';
@@ -11,14 +12,31 @@ import {
   IconFilledCircle,
   IconOutlinedArrowBack,
   IconOutlinedArrowForward,
+  IconOutlinedSearch,
   IconOutlinedTrendingFlat,
 } from '../../icons';
 import {TextBlock} from '../../text-block';
 import {createCustomThemeWithBaseThemeSwitch} from '../../test/theme-select-object';
-import {CreateThemeArgs, ThemeProvider} from '../../theme';
-import {StyledButton} from '../styled';
+import {Select, SelectOption, SelectOptionProps} from '../../select';
+import {
+  PaginationItemProps,
+  PaginationItemsProps,
+  PaginationLayoutItem,
+} from '../types';
+import {getPaginationItemsLayout} from '../utils';
+import {usePaginationContext} from '../context';
+import {Form, FormInput} from '../../form';
+import {IconButton} from '../../icon-button';
+import {TextField} from '../../text-field';
+import {GridLayout} from '../../grid-layout';
+import {PaginationListItem} from '../components/list-item';
+import {PaginationButton} from '../components/button';
+import {ButtonOrButtonLinkProps} from '../../button';
+import {useReactKeys} from '../../utils/hooks';
+import {filterOutFalsyProperties} from '../../utils/filter-object';
+import {ScreenReaderOnly} from '../../screen-reader-only/screen-reader-only';
 
-const paginationItemCustomThemeObject: CreateThemeArgs = {
+const paginationItemCustomThemeObject = {
   name: 'pagination-static-theme',
   overrides: {
     stylePresets: {
@@ -28,16 +46,70 @@ const paginationItemCustomThemeObject: CreateThemeArgs = {
           color: '{{colors.interfaceInformative010}}',
         },
       },
+      paginationItemInput: {
+        base: {
+          backgroundColor: '{{colors.transparent}}',
+          color: '{{colors.interactiveLink010}}',
+          iconColor: '{{colors.interactiveLink010}}',
+          cursor: 'default',
+        },
+        disabled: {
+          color: '{{colors.inkNonEssential}}',
+          iconColor: '{{colors.inkNonEssential}}',
+        },
+        'focus-visible': {
+          outlineColor: '{{outlines.outlineColorDefault}}',
+          outlineStyle: '{{outlines.outlineStyleDefault}}',
+          outlineWidth: '{{outlines.outlineWidthDefault}}',
+          safariOutlineStyle: '{{outlines.safariOutlineStyleDefault}}',
+        },
+      },
+    },
+  },
+} as CreateThemeArgs;
+
+const customDefaults = {
+  paginationItemInput: {
+    small: {
+      typographyPreset: 'utilityButton010',
+      stylePreset: 'paginationItemInput',
+      iconSize: 'iconSize010',
+      paddingInline: 'space020',
+      marginInline: 'space010',
+      marginBlockEnd: 'space000',
+    },
+    medium: {
+      typographyPreset: 'utilityButton020',
+      stylePreset: 'paginationItemInput',
+      iconSize: 'iconSize020',
+      paddingInline: 'space020',
+      marginInline: 'space010',
+      marginBlockEnd: 'space000',
+    },
+    large: {
+      typographyPreset: 'utilityButton030',
+      stylePreset: 'paginationItemInput',
+      iconSize: 'iconSize020',
+      paddingInline: 'space020',
+      marginInline: 'space010',
+      marginBlockEnd: 'space000',
     },
   },
 };
 
-const defaultProps = {
+// eslint-disable-next-line
+const onPageChange = (page: number) => console.log(`Page changed to ${page}`);
+
+const uncontrolledProps = {
   totalItems: 80,
   pageSize: 10,
   defaultPage: 4,
-  // eslint-disable-next-line
-  onPageChange: (page: number) => console.log(`Page changed to ${page}`),
+  onPageChange,
+};
+const controlledProps = {
+  totalItems: 232,
+  pageSize: 10,
+  page: 4,
 };
 // defaults to truncation: true, siblings: 3, boundaries: 1
 const defaultItemsProps = {truncation: false, siblings: 0, boundaries: 0};
@@ -46,7 +118,7 @@ export const StoryDefault = () => (
   <StorybookPage columns={{md: 'auto'}}>
     <StorybookCase title="HREF-based">
       <Pagination
-        {...defaultProps}
+        {...controlledProps}
         buildHref={(page: number) => `page/${page}`}
         aria-label="href"
       >
@@ -58,7 +130,7 @@ export const StoryDefault = () => (
       </Pagination>
     </StorybookCase>
     <StorybookCase title="onclick-based">
-      <Pagination {...defaultProps} aria-label="on-click">
+      <Pagination {...uncontrolledProps} aria-label="on-click">
         <PaginationFirstItem />
         <PaginationPrevItem />
         <PaginationItems {...defaultItemsProps} />
@@ -73,7 +145,7 @@ StoryDefault.storyName = 'Default';
 export const StorySize = () => (
   <StorybookPage columns={{md: 'auto'}}>
     <StorybookCase title="Small">
-      <Pagination {...defaultProps} size="small" aria-label="small">
+      <Pagination {...uncontrolledProps} size="small" aria-label="small">
         <PaginationFirstItem />
         <PaginationPrevItem />
         <PaginationItems {...defaultItemsProps} />
@@ -82,7 +154,7 @@ export const StorySize = () => (
       </Pagination>
     </StorybookCase>
     <StorybookCase title="Medium">
-      <Pagination {...defaultProps} aria-label="medium">
+      <Pagination {...uncontrolledProps} aria-label="medium">
         <PaginationFirstItem />
         <PaginationPrevItem />
         <PaginationItems {...defaultItemsProps} />
@@ -91,7 +163,7 @@ export const StorySize = () => (
       </Pagination>
     </StorybookCase>
     <StorybookCase title="Large">
-      <Pagination {...defaultProps} size="large" aria-label="large">
+      <Pagination {...uncontrolledProps} size="large" aria-label="large">
         <PaginationFirstItem />
         <PaginationPrevItem />
         <PaginationItems {...defaultItemsProps} />
@@ -106,18 +178,18 @@ StorySize.storyName = 'Size';
 export const StoryVariationsInNavigation = () => (
   <StorybookPage columns={{md: 'auto'}}>
     <StorybookCase title="Numbers only">
-      <Pagination {...defaultProps} aria-label="numbers only">
+      <Pagination {...uncontrolledProps} aria-label="numbers only">
         <PaginationItems {...defaultItemsProps} />
       </Pagination>
     </StorybookCase>
     <StorybookCase title="Next and previous only">
-      <Pagination {...defaultProps} aria-label="next and previous only">
+      <Pagination {...uncontrolledProps} aria-label="next and previous only">
         <PaginationPrevItem />
         <PaginationNextItem />
       </Pagination>
     </StorybookCase>
     <StorybookCase title="Next and previous labels">
-      <Pagination {...defaultProps} aria-label="next and previous labels">
+      <Pagination {...uncontrolledProps} aria-label="next and previous labels">
         <PaginationPrevItem overrides={{width: 'auto'}}>
           Previous
         </PaginationPrevItem>
@@ -128,7 +200,7 @@ export const StoryVariationsInNavigation = () => (
       </Pagination>
     </StorybookCase>
     <StorybookCase title="First and last labels">
-      <Pagination {...defaultProps} aria-label="first and last labels">
+      <Pagination {...uncontrolledProps} aria-label="first and last labels">
         <PaginationFirstItem overrides={{width: 'auto'}}>
           First
         </PaginationFirstItem>
@@ -149,6 +221,7 @@ export const StoryVariationsInTruncation = () => (
         totalItems={232}
         pageSize={10}
         defaultPage={9}
+        onPageChange={onPageChange}
         aria-label="no truncation of a long list"
       >
         <PaginationItems truncation={false} />
@@ -159,7 +232,8 @@ export const StoryVariationsInTruncation = () => (
         totalItems={232}
         pageSize={10}
         defaultPage={9}
-        aria-label="base"
+        onPageChange={onPageChange}
+        aria-label="siblings 3, boundaries 1"
       >
         <PaginationItems />
       </Pagination>
@@ -169,7 +243,8 @@ export const StoryVariationsInTruncation = () => (
         totalItems={232}
         pageSize={10}
         defaultPage={9}
-        aria-label="base"
+        onPageChange={onPageChange}
+        aria-label="siblings 3, boundaries 0"
       >
         <PaginationFirstItem />
         <PaginationItems siblings={3} boundaries={0} />
@@ -181,7 +256,8 @@ export const StoryVariationsInTruncation = () => (
         totalItems={232}
         pageSize={10}
         defaultPage={9}
-        aria-label="base"
+        onPageChange={onPageChange}
+        aria-label="siblings 3, boundaries 2"
       >
         <PaginationItems siblings={2} boundaries={2} />
       </Pagination>
@@ -190,45 +266,489 @@ export const StoryVariationsInTruncation = () => (
 );
 StoryVariationsInTruncation.storyName = 'Variations in truncation';
 
-export const StoryVariationsInInput = () => (
-  <StorybookPage columns={{md: 'auto'}}>
-    <StorybookCase title="Display only">
-      <Pagination
-        totalItems={232}
-        pageSize={10}
-        defaultPage={9}
-        aria-label="display only"
+export const StoryVariationsInInput = () => {
+  const CustomInputPaginationItemButton = ({
+    size = 'medium',
+    pageNumber,
+    lastPage = 1,
+    changePage = () => {},
+  }: ButtonOrButtonLinkProps & PaginationItemProps) => {
+    const inputRef = useRef<HTMLInputElement>(null);
+    const validateAndChangePage = (newPage: number) => {
+      if (newPage >= 1 && newPage <= lastPage) {
+        changePage(newPage);
+      }
+      inputRef.current?.blur();
+    };
+
+    const ariaId = useReactKeys(1)[0];
+    const inputLabelId = `page-input-label${ariaId}`;
+    const inputId = `page-input${ariaId}`;
+
+    // Cannot use a stylePreset (such as paginationItemInput) on the TextField or it will lose the default border
+    const {
+      paddingInline,
+      marginInline,
+      marginBlockEnd,
+    } = customDefaults.paginationItemInput[size];
+
+    return (
+      <Form
+        onSubmit={values => validateAndChangePage(parseInt(values.newPage, 10))}
       >
-        <PaginationFirstItem />
-        <PaginationPrevItem />
-        <PaginationItems
-          truncation
-          siblings={0}
-          boundaries={0}
-          overrides={{
-            itemButton: ({pageNumber, lastPage}) => (
-              <TextBlock
-                paddingInline="space020"
-                typographyPreset="utilityButton020"
-                stylePreset="paginationItemNonInteractive"
+        <FormInput name="newPage">
+          <ScreenReaderOnly id={inputLabelId} aria-hidden="true">
+            Enter page between 1 and {lastPage}
+          </ScreenReaderOnly>
+          <TextField
+            id={inputId}
+            aria-labelledby={inputLabelId}
+            ref={inputRef}
+            size="small"
+            defaultValue={pageNumber}
+            onBlur={props =>
+              validateAndChangePage(parseInt(props.target?.value, 10))
+            }
+            endEnhancer={
+              <IconButton
+                type="submit"
+                title="Change Page"
+                overrides={{
+                  stylePreset: 'paginationItemInput',
+                }}
               >
-                Page {pageNumber} of {lastPage}
-              </TextBlock>
-            ),
-          }}
-        />
-        <PaginationNextItem />
-        <PaginationLastItem />
-      </Pagination>
-    </StorybookCase>
-  </StorybookPage>
-);
+                <IconOutlinedSearch />
+              </IconButton>
+            }
+            overrides={{
+              paddingInline,
+              marginInline,
+              marginBlockEnd,
+              width: '84px',
+              endEnhancer: {
+                spaceInline: 'space010',
+              },
+            }}
+          />
+        </FormInput>
+      </Form>
+    );
+  };
+
+  const CustomInputPaginationItem = React.forwardRef<
+    HTMLElement,
+    PaginationItemProps
+  >(({pageNumber, overrides = {}}, ref) => {
+    const {size = 'medium', page, lastPage, ...rest} = usePaginationContext();
+    const {itemDescription} = overrides;
+    let ItemDescription = null;
+    let textblockSettings: typeof overrides = {};
+    const theme = useTheme();
+
+    if (itemDescription) {
+      ItemDescription = itemDescription;
+      textblockSettings = {
+        ...theme.componentDefaults.paginationItemNonInteractive[size],
+        ...filterOutFalsyProperties(itemDescription),
+      };
+    }
+
+    return (
+      <PaginationListItem key="pages" size={size}>
+        <GridLayout
+          columns="auto auto"
+          rows="1fr"
+          autoFlow="row"
+          alignItems="center"
+          columnGap="space010"
+        >
+          <CustomInputPaginationItemButton
+            ref={ref}
+            pageNumber={page}
+            selected
+            {...rest}
+          />
+          {itemDescription && (
+            // @ts-ignore
+            <ItemDescription
+              {...textblockSettings}
+              selected={pageNumber === page}
+              pageNumber={page}
+              lastPage={lastPage}
+            />
+          )}
+        </GridLayout>
+      </PaginationListItem>
+    );
+  });
+
+  return (
+    <>
+      <StorybookPage>
+        <StorybookCase title="Description only">
+          <Pagination
+            totalItems={232}
+            pageSize={10}
+            defaultPage={9}
+            onPageChange={onPageChange}
+            aria-label="description only"
+          >
+            <PaginationFirstItem />
+            <PaginationPrevItem />
+            <PaginationItems
+              truncation
+              siblings={0}
+              boundaries={0}
+              overrides={{
+                itemButton: () => null,
+                itemDescription: ({
+                  pageNumber,
+                  lastPage,
+                  ...textBlockStyles
+                }) => (
+                  <TextBlock {...textBlockStyles}>
+                    Page {pageNumber} of {lastPage}
+                  </TextBlock>
+                ),
+              }}
+            />
+            <PaginationNextItem />
+            <PaginationLastItem />
+          </Pagination>
+        </StorybookCase>
+        <StorybookCase title="Input only">
+          <Pagination
+            totalItems={232}
+            pageSize={10}
+            defaultPage={9}
+            onPageChange={onPageChange}
+            aria-label="input only"
+          >
+            <PaginationFirstItem />
+            <PaginationPrevItem />
+            <PaginationItems
+              truncation
+              siblings={0}
+              boundaries={0}
+              overrides={{
+                itemButton: props => (
+                  <CustomInputPaginationItemButton {...props} />
+                ),
+              }}
+            />
+            <PaginationNextItem />
+            <PaginationLastItem />
+          </Pagination>
+        </StorybookCase>
+        <StorybookCase title="Input and description">
+          <Pagination
+            totalItems={232}
+            pageSize={10}
+            defaultPage={9}
+            onPageChange={onPageChange}
+            aria-label="input and description"
+          >
+            <PaginationFirstItem />
+            <PaginationPrevItem />
+            <PaginationItems
+              truncation
+              siblings={0}
+              boundaries={0}
+              overrides={{
+                itemButton: props => (
+                  <CustomInputPaginationItemButton {...props} />
+                ),
+                itemDescription: ({lastPage, ...textBlockStyles}) => (
+                  <TextBlock {...textBlockStyles}>
+                    of {lastPage} pages
+                  </TextBlock>
+                ),
+              }}
+            />
+            <PaginationNextItem />
+            <PaginationLastItem />
+          </Pagination>
+        </StorybookCase>
+      </StorybookPage>
+      <StorybookPage columns="1fr">
+        <StorybookCase title="Buttons with Input at end">
+          <Pagination
+            totalItems={232}
+            pageSize={10}
+            defaultPage={9}
+            onPageChange={onPageChange}
+            aria-label="input at end"
+          >
+            <PaginationFirstItem />
+            <PaginationPrevItem />
+            <PaginationItems truncation siblings={3} boundaries={1} />
+            <PaginationNextItem />
+            <PaginationLastItem />
+            <CustomInputPaginationItem overrides={{}} />
+          </Pagination>
+        </StorybookCase>
+        <StorybookCase title="Buttons with Input for selected item">
+          <Pagination
+            totalItems={232}
+            pageSize={10}
+            defaultPage={9}
+            onPageChange={onPageChange}
+            aria-label="input for selected item"
+          >
+            <PaginationFirstItem />
+            <PaginationPrevItem />
+            <PaginationItems
+              truncation
+              siblings={3}
+              boundaries={1}
+              overrides={{
+                itemButton: ({selected, ...rest}) => {
+                  if (selected) {
+                    return (
+                      <CustomInputPaginationItemButton
+                        selected={selected}
+                        {...rest}
+                      />
+                    );
+                  }
+
+                  return (
+                    <PaginationButton {...rest}>
+                      {rest.pageNumber}
+                    </PaginationButton>
+                  );
+                },
+              }}
+            />
+            <PaginationNextItem />
+            <PaginationLastItem />
+          </Pagination>
+        </StorybookCase>
+      </StorybookPage>
+    </>
+  );
+};
 StoryVariationsInInput.storyName = 'Variations in input';
 
+export const StoryVariationsInSelection = () => {
+  const CustomDropdownPaginationItems = ({
+    children,
+    overrides = {},
+    truncation = true,
+    siblings = 3,
+    boundaries = 1,
+    eventContext = {},
+    eventOriginator = 'pagination-dropdown',
+  }: PaginationItemsProps) => {
+    const theme = useTheme();
+    const {
+      size = 'medium',
+      buildHref,
+      /* istanbul ignore next */
+      changePage = () => {},
+      page,
+      lastPage,
+    } = usePaginationContext();
+
+    const {itemDescription, ...logicalOverrides} = overrides;
+    let ItemDescription = null;
+    let textblockSettings: typeof overrides = {};
+
+    if (itemDescription) {
+      ItemDescription = itemDescription;
+      textblockSettings = {
+        ...theme.componentDefaults.paginationItemNonInteractive[size],
+        ...filterOutFalsyProperties(itemDescription),
+      };
+    }
+
+    const paginationElements = [] as React.ReactElement<SelectOptionProps>[];
+    if (page && lastPage) {
+      const layout: [PaginationLayoutItem?] = getPaginationItemsLayout({
+        page,
+        lastPage,
+        truncation,
+        siblings,
+        boundaries,
+      });
+
+      let truncationCount = 0;
+      /* istanbul ignore next */
+      layout.forEach((element: PaginationLayoutItem = 1) => {
+        switch (element) {
+          case '-':
+            truncationCount += 1;
+            paginationElements.push(
+              <SelectOption
+                key={`trunc${truncationCount}`}
+                value=""
+                overrides={logicalOverrides}
+              >
+                ...
+              </SelectOption>,
+            );
+            break;
+          default:
+            {
+              const pageNumber: number = element;
+              paginationElements.push(
+                <SelectOption
+                  key={`page${pageNumber}`}
+                  value={pageNumber.toString()}
+                  selected={pageNumber === page}
+                  overrides={logicalOverrides}
+                >
+                  {children}
+                  {pageNumber}
+                </SelectOption>,
+              );
+            }
+            break;
+        }
+      });
+    }
+
+    return (
+      <PaginationListItem key="pages" size={size}>
+        <GridLayout
+          columns="auto auto"
+          rows="1fr"
+          autoFlow="row"
+          alignItems="center"
+          columnGap="space030"
+        >
+          <Select
+            size="small"
+            eventContext={eventContext}
+            eventOriginator={eventOriginator}
+            overrides={{
+              button: {
+                marginInline: 'space010',
+                marginBlockEnd: 'space000',
+              },
+            }}
+            onChange={event => {
+              if (event.target.value) {
+                const pageNumber = Number(event.target.value);
+                changePage(pageNumber);
+                const href = buildHref! && buildHref(pageNumber);
+                if (href) {
+                  window.location.href = href;
+                }
+              }
+            }}
+          >
+            {paginationElements as []}
+          </Select>
+          {itemDescription && (
+            // @ts-ignore
+            <ItemDescription
+              {...textblockSettings}
+              selected
+              pageNumber={page}
+              lastPage={lastPage}
+            />
+          )}
+        </GridLayout>
+      </PaginationListItem>
+    );
+  };
+
+  return (
+    <StorybookPage>
+      <StorybookCase title="Dropdown (HREF-based)">
+        <Pagination
+          totalItems={232}
+          pageSize={10}
+          defaultPage={9}
+          buildHref={(page: number) => `page/${page}`}
+          aria-label="dropdown href based"
+        >
+          <PaginationFirstItem />
+          <PaginationPrevItem />
+          <CustomDropdownPaginationItems truncation siblings={3} boundaries={1}>
+            Page{' '}
+          </CustomDropdownPaginationItems>
+          <PaginationNextItem />
+          <PaginationLastItem />
+        </Pagination>
+      </StorybookCase>
+      <StorybookCase title="Dropdown (onclick-based)">
+        <Pagination
+          totalItems={232}
+          pageSize={10}
+          defaultPage={9}
+          onPageChange={onPageChange}
+          aria-label="dropdown onclick based"
+        >
+          <PaginationFirstItem />
+          <PaginationPrevItem />
+          <CustomDropdownPaginationItems truncation siblings={3} boundaries={1}>
+            Page{' '}
+          </CustomDropdownPaginationItems>
+          <PaginationNextItem />
+          <PaginationLastItem />
+        </Pagination>
+      </StorybookCase>
+      <StorybookCase title="Dropdown and short description">
+        <Pagination
+          totalItems={232}
+          pageSize={10}
+          defaultPage={9}
+          onPageChange={onPageChange}
+          aria-label="dropdown and short description"
+        >
+          <PaginationFirstItem />
+          <PaginationPrevItem />
+          <CustomDropdownPaginationItems
+            truncation
+            siblings={3}
+            boundaries={1}
+            overrides={{
+              itemDescription: ({lastPage, ...textBlockStyles}) => (
+                <TextBlock {...textBlockStyles}>of {lastPage}</TextBlock>
+              ),
+            }}
+          >
+            Page{' '}
+          </CustomDropdownPaginationItems>
+          <PaginationNextItem />
+          <PaginationLastItem />
+        </Pagination>
+      </StorybookCase>
+      <StorybookCase title="Dropdown and long description">
+        <Pagination
+          totalItems={232}
+          pageSize={10}
+          defaultPage={9}
+          onPageChange={onPageChange}
+          aria-label="dropdown and long description"
+        >
+          <PaginationFirstItem />
+          <PaginationPrevItem />
+          <CustomDropdownPaginationItems
+            truncation
+            siblings={3}
+            boundaries={1}
+            overrides={{
+              itemDescription: ({lastPage, ...textBlockStyles}) => (
+                <TextBlock {...textBlockStyles}>of {lastPage} pages</TextBlock>
+              ),
+            }}
+          />
+          <PaginationNextItem />
+          <PaginationLastItem />
+        </Pagination>
+      </StorybookCase>
+    </StorybookPage>
+  );
+};
+
+StoryVariationsInSelection.storyName = 'Variations in selection';
 export const StoryLogicalProps = () => (
   <StorybookPage columns={{md: 'auto'}}>
     <StorybookCase title="Custom button size">
-      <Pagination {...defaultProps} aria-label="custom button size">
+      <Pagination {...uncontrolledProps} aria-label="custom button size">
         <PaginationPrevItem />
         <PaginationItems
           {...defaultItemsProps}
@@ -238,7 +758,7 @@ export const StoryLogicalProps = () => (
       </Pagination>
     </StorybookCase>
     <StorybookCase title="Logical props - items padding">
-      <Pagination {...defaultProps} aria-label="items padding overrides">
+      <Pagination {...uncontrolledProps} aria-label="items padding overrides">
         <PaginationPrevItem />
         <PaginationItems
           {...defaultItemsProps}
@@ -248,7 +768,7 @@ export const StoryLogicalProps = () => (
       </Pagination>
     </StorybookCase>
     <StorybookCase title="Logical props - items margin">
-      <Pagination {...defaultProps} aria-label="items margin overrides">
+      <Pagination {...uncontrolledProps} aria-label="items margin overrides">
         <PaginationPrevItem />
         <PaginationItems
           {...defaultItemsProps}
@@ -259,7 +779,7 @@ export const StoryLogicalProps = () => (
     </StorybookCase>
     <StorybookCase title="Logical props - component margin">
       <Pagination
-        {...defaultProps}
+        {...uncontrolledProps}
         aria-label="component margin overrides"
         overrides={{marginInline: 'space060'}}
       >
@@ -275,7 +795,7 @@ StoryLogicalProps.storyName = 'Logical props';
 export const StoryStyleOverrides = () => (
   <StorybookPage columns={{md: 'auto'}}>
     <StorybookCase title="Custom colour">
-      <Pagination {...defaultProps} aria-label="custom colour">
+      <Pagination {...uncontrolledProps} aria-label="custom colour">
         <PaginationPrevItem />
         <PaginationItems
           {...defaultItemsProps}
@@ -287,7 +807,7 @@ export const StoryStyleOverrides = () => (
       </Pagination>
     </StorybookCase>
     <StorybookCase title="Custom style">
-      <Pagination {...defaultProps} aria-label="custom style">
+      <Pagination {...uncontrolledProps} aria-label="custom style">
         <PaginationPrevItem />
         <PaginationItems
           {...defaultItemsProps}
@@ -302,7 +822,10 @@ export const StoryStyleOverrides = () => (
       </Pagination>
     </StorybookCase>
     <StorybookCase title="Custom previous and next icons">
-      <Pagination {...defaultProps} aria-label="custom previous and next icons">
+      <Pagination
+        {...uncontrolledProps}
+        aria-label="custom previous and next icons"
+      >
         <PaginationPrevItem
           overrides={{
             icon: () => (
@@ -337,6 +860,7 @@ export const StoryStyleOverrides = () => (
         totalItems={232}
         pageSize={10}
         defaultPage={9}
+        onPageChange={onPageChange}
         aria-label="truncation with icon"
       >
         <PaginationPrevItem />
@@ -364,7 +888,7 @@ export const StoryNumberOverrides = () => (
   <StorybookPage columns={{md: 'auto'}}>
     <StorybookCase title="Items as clickable icon buttons (using children)">
       <Pagination
-        {...defaultProps}
+        {...uncontrolledProps}
         aria-label="items as clickable icon buttons (using children)"
       >
         <PaginationPrevItem />
@@ -376,7 +900,7 @@ export const StoryNumberOverrides = () => (
     </StorybookCase>
     <StorybookCase title="Items as clickable icon buttons (using overrides.itemButton)">
       <Pagination
-        {...defaultProps}
+        {...uncontrolledProps}
         aria-label="items as clickable icons (using overrides.itemButton)"
       >
         <PaginationPrevItem />
@@ -384,9 +908,9 @@ export const StoryNumberOverrides = () => (
           truncation={false}
           overrides={{
             itemButton: props => (
-              <StyledButton {...props}>
+              <PaginationButton {...props}>
                 <IconFilledCircle />
-              </StyledButton>
+              </PaginationButton>
             ),
           }}
         />
@@ -395,7 +919,7 @@ export const StoryNumberOverrides = () => (
     </StorybookCase>
     <StorybookCase title="Items as unclickable icons (using overrides.itemButton)">
       <Pagination
-        {...defaultProps}
+        {...uncontrolledProps}
         aria-label="items as unclickable icons (using overrides.itemButton)"
       >
         <PaginationPrevItem />
