@@ -38,6 +38,7 @@ import {
 import {TextBlock} from '../../text-block';
 import {EventTrigger} from '../../instrumentation';
 import {compileTheme, createTheme} from '../../theme';
+import {PaginationListItem} from '../components/list-item';
 
 const paginationItemContent = 'Pagination Item';
 const textBlock = '...';
@@ -101,10 +102,16 @@ type PaginationWithItemProps = {
   itemProps: PaginationItemProps;
 };
 
+// NB This approach is not recommended in real life. PaginationListItems is the public API
+// of generating PaginationListItem and PaginationItem components automatically.
 const PaginationWithItem = ({props, itemProps}: PaginationWithItemProps) => (
-  <Pagination {...props}>
-    <PaginationItem size="small" key="1" {...itemProps} />
-  </Pagination>
+  <PaginationListItem>
+    <Pagination {...props}>
+      <PaginationListItem>
+        <PaginationItem size="small" key="1" {...itemProps} />
+      </PaginationListItem>
+    </Pagination>
+  </PaginationListItem>
 );
 
 describe('Pagination and PaginationItems only', () => {
@@ -124,6 +131,7 @@ describe('Pagination and PaginationItems only', () => {
       pageSize: 10,
       defaultPage: 4,
       buildHref,
+      children: null,
     };
     controlledProps = {
       totalItems: 232,
@@ -131,6 +139,7 @@ describe('Pagination and PaginationItems only', () => {
       page: 4,
       defaultPage: 4,
       onPageChange,
+      children: null,
     };
   });
 
@@ -482,69 +491,25 @@ describe('Pagination and PaginationItems only', () => {
       expect(fragment).toMatchSnapshot();
     });
 
-    test('fire tracking event on normal item with href set and onPageChange unset', async () => {
-      const mockFireEvent = jest.fn();
+    it('renders with overrides.itemDescription', () => {
       const itemProps = {
+        size: 'medium' as PaginationSize,
+        children: paginationItemContent,
         ...pageItemProps,
-        lastPage,
-        children: paginationItemContent,
-        key: '1',
-        eventOriginator: 'pagination-item',
-        eventContext: {
-          event: 'event data',
+        overrides: {
+          itemButton: () => null,
+          itemDescription: props => (
+            <span>
+              Page {props.pageNumber} of {props.lastPage}
+            </span>
+          ),
         },
-      };
-
-      const {getByTestId} = renderWithImplementation(
-        PaginationWithItem,
-        {props: defaultProps, itemProps},
-        mockFireEvent,
-      );
-      const paginationItemButton = getByTestId('pagination-item');
-      fireEvent.click(paginationItemButton);
-      expect(mockFireEvent).toHaveBeenCalledWith({
-        originator: 'pagination-item',
-        trigger: EventTrigger.Click,
-        context: {
-          href: 'https://paginationitem.test',
-          event: 'event data',
-        },
+      } as PaginationItemProps;
+      const fragment = renderToFragmentWithTheme(PaginationWithItem, {
+        props: defaultProps,
+        itemProps,
       });
-      expect(onPageChange).not.toHaveBeenCalled();
-    });
-
-    test('fire tracking event on normal item with href unset and onPageChange set', async () => {
-      jest.spyOn(console, 'warn').mockImplementation();
-      const mockFireEvent = jest.fn();
-      const itemProps = {
-        selected,
-        pageNumber,
-        onClick: () => onPageChange(pageNumber),
-        lastPage,
-        children: paginationItemContent,
-        key: '1',
-        eventOriginator: 'pagination-item',
-        eventContext: {
-          event: 'event data',
-        },
-      };
-
-      const {getByTestId} = renderWithImplementation(
-        PaginationWithItem,
-        {props: controlledProps, itemProps},
-        mockFireEvent,
-      );
-      const paginationItemButton = getByTestId('pagination-item');
-      fireEvent.click(paginationItemButton);
-      expect(mockFireEvent).toHaveBeenCalledWith({
-        originator: 'pagination-item',
-        trigger: EventTrigger.Click,
-        context: {
-          page: pageNumber,
-          event: 'event data',
-        },
-      });
-      expect(onPageChange).toHaveBeenCalledWith(pageNumber);
+      expect(fragment).toMatchSnapshot();
     });
 
     it('renders selected pagination item with aria attributes', () => {
