@@ -1,8 +1,8 @@
 import React from 'react';
-import {useTheme} from '@emotion/react';
+import {useTheme} from '../../../theme';
 import {IconFilledFirstPage, NewsKitIconProps} from '../../../icons';
 import {usePaginationContext} from '../../context';
-import {StyledListItem} from '../../styled';
+import {PaginationListItem} from '../list-item';
 import {PaginationItem} from '../item/pagination-item';
 import {PaginationFirstItemProps} from './types';
 import {PaginationItemType} from '../../types';
@@ -42,13 +42,14 @@ export const PaginationFirstItem = React.forwardRef<
 
   const href = buildHref! && buildHref(1);
   return (
-    <StyledListItem key="first">
+    <PaginationListItem key="first">
       <PaginationItem
         // @ts-ignore
         itemType={itemType}
         data-testid="pagination-first-item"
         eventOriginator="pagination-first-item"
         eventContext={eventContext}
+        pageNumber={1}
         {...propsFromContext}
         overrides={overrides}
         ref={ref}
@@ -58,6 +59,6 @@ export const PaginationFirstItem = React.forwardRef<
         <PaginationIcon />
         {children}
       </PaginationItem>
-    </StyledListItem>
+    </PaginationListItem>
   );
 });
