@@ -1,8 +1,8 @@
 import React from 'react';
-import {useTheme} from '@emotion/react';
+import {useTheme} from '../../../theme';
 import {IconFilledChevronLeft, NewsKitIconProps} from '../../../icons';
 import {usePaginationContext} from '../../context';
-import {StyledListItem} from '../../styled';
+import {PaginationListItem} from '../list-item';
 import {PaginationItem} from '../item/pagination-item';
 import {PaginationPrevItemProps} from './types';
 import {filterOutFalsyProperties} from '../../../utils/filter-object';
@@ -45,13 +45,14 @@ export const PaginationPrevItem = React.forwardRef<
   const prevPage = Math.max(page - 1, 1);
   const href = buildHref! && buildHref(prevPage);
   return (
-    <StyledListItem key="prev">
+    <PaginationListItem key="prev">
       <PaginationItem
         // @ts-ignore
         itemType={itemType}
         data-testid="pagination-prev-item"
         eventOriginator="pagination-prev-item"
         eventContext={eventContext}
+        pageNumber={prevPage}
         {...propsFromContext}
         overrides={overrides}
         ref={ref}
@@ -61,6 +62,6 @@ export const PaginationPrevItem = React.forwardRef<
         <PaginationIcon />
         {children}
       </PaginationItem>
-    </StyledListItem>
+    </PaginationListItem>
   );
 });
