@@ -1,9 +1,12 @@
+export * from './components/list-item';
+export * from './components/items';
 export * from './components/first-item';
 export * from './components/last-item';
+export * from './components/item';
+export * from './components/button';
 export * from './components/next-item';
 export * from './components/prev-item';
-export * from './components/item';
-export * from './components/items';
 export * from './context';
 export * from './pagination';
 export * from './types';
+export * from './utils';
