@@ -2,7 +2,7 @@ import {MQ} from '../utils';
 import {LogicalProps} from '../utils/logical-properties';
 
 import {EventData} from '../instrumentation';
-import {Override} from '../utils/overrides';
+import {ComponentOverrides, Override} from '../utils/overrides';
 import {NewsKitIconProps} from '../icons';
 import {ButtonOrButtonLinkProps, ButtonProps} from '../button/types';
 
@@ -11,6 +11,7 @@ import {PaginationFirstItemProps} from './components/first-item/types';
 import {PaginationLastItemProps} from './components/last-item/types';
 import {PaginationNextItemProps} from './components/next-item/types';
 import {PaginationPrevItemProps} from './components/prev-item/types';
+import {TextBlockProps} from '../text-block';
 
 export type PaginationSize = 'small' | 'medium' | 'large';
 
@@ -40,7 +41,7 @@ export interface PaginationItemAriaInput {
 }
 
 export interface PaginationProps extends React.AriaAttributes {
-  children?: Exclude<React.ReactNode, 'undefined'>;
+  children: Exclude<React.ReactNode, 'undefined'>;
   size?: PaginationSize;
   pageSize: number;
   page?: number;
@@ -50,40 +51,56 @@ export interface PaginationProps extends React.AriaAttributes {
   onPageChange?: (pageNumber: number) => void;
   overrides?: {
     stylePreset?: MQ<string>;
+    typographyPreset?: MQ<string>;
   } & LogicalProps;
 }
 
 export interface ComponentSizeProps extends React.AriaAttributes {
+  children?: React.ReactNode;
   path?: string;
   size?: PaginationSize;
 }
 
+export interface PaginationItemDescriptionProps extends ComponentOverrides {
+  selected?: boolean;
+  pageNumber?: number;
+  lastPage?: number;
+}
+
 export interface PaginationItemsProps extends ButtonProps, EventData {
-  children?: Exclude<React.ReactNode, 'undefined'>;
+  children?: React.ReactNode;
   truncation?: boolean;
   siblings?: number;
   boundaries?: number;
   overrides?: ButtonProps['overrides'] & {
     icon?: Override<NewsKitIconProps>;
     itemButton?: Override<ButtonOrButtonLinkProps & PaginationItemProps>;
+    itemDescription?: Override<TextBlockProps & PaginationItemDescriptionProps>;
   };
 }
 
-export interface PaginationItemProps extends ButtonProps, EventData {
-  children?: Exclude<React.ReactNode, 'undefined'>;
+export interface PaginationItemProps extends EventData {
+  children?: React.ReactNode;
+  ref?: React.RefObject<HTMLElement> | React.ForwardedRef<HTMLElement>;
   itemType?: PaginationItemType;
-  selected?: boolean;
-  lastPage: number;
   pageNumber?: number;
   pageSize?: number;
   totalItems?: number;
   href?: string;
+  // The below 3 are derived props
+  selected?: boolean;
+  lastPage?: number;
+  changePage?: (pageNumber: number) => void;
+  // These render functions can be used to change the appearance of the
+  // truncation icon, the number button and the text description, respectively
   overrides?: ButtonProps['overrides'] & {
     icon?: Override<NewsKitIconProps>;
     itemButton?: Override<ButtonOrButtonLinkProps & PaginationItemProps>;
+    itemDescription?: Override<TextBlockProps & PaginationItemDescriptionProps>;
   };
 }
 
+// The usePaginationContext hook returns this data
 export interface PaginationProviderContext {
   size: PaginationSize;
   pageSize: number;
